Extract chordpro bits into separate job
